package tree-sitter-ocaml
tree-sitter-ocaml is interesting because it has two grammars in
one repository, ocaml in ocaml/ and ocaml-interface in interface/.
We add some optional pieces to 'tspm.grammar' which allows us to
include a set of paths within the source repository and use one
path as the primary path to compile and include in the package.

We also make the change that we no longer separate out the src/
directory. Instead, the 'src' packages are a copy of everything
generated into '$subpath/src' plus any licenses and notices.
Previously the behavior of 'src' was to have a src/ directory
which contained the generated files and also top-level were the
licenses and notices.