strip dependencies and customize flake systems
Here we strip out all other flake dependencies and use similar
functions to edolstra/dwarffs and numtide/flake-utils to build
a custom flake output that takes system into account. It appears
from my bug hunting that the 'nix flake check' failure was introduced
with '{ path = "metadata"; }', which uses `builtins.hashFile` on
a built derivation, which, if I understand it correctly, is
recursive Nix. That appears to mess with the system that
builtins.hashFile needs somehow, so the builder for aarch64-linux
fails.