fix build for tree-sitter-ocaml-interface
The common/ directory was not being copied over. 'includePaths' has
been removed and in its place now is 'copyPaths' which straight up
copies directories into 'tsDir'. This allows us to copy over files
we need that don't follow a generic format but also be stringent
about the main grammar files.