replace 'author' with 'owner' nomenclature
Author is fine most of the time but it falls apart when it comes
to forks. The forker is the owner of the repository but not
necessarily the author of the grammar - they might have no commits
to grammar.js itself. Similarly, what if a grammar is transferred?
If a grammar gets moved to 'tree-sitter', the "author" has changed -
does that make sense? I think "owner" is a blanket better term for
the relationship TSPM is describing.